turned cpi marker publisher into its own plugin
