+++ src/giskardpy/plugin_visualization.py
@@ -8,23 +8,15 @@ from visualization_msgs.msg import Marker, MarkerArray
 
 from giskardpy.tfwrapper import pose_to_kdl, kdl_to_pose
 from plugin import GiskardBehavior
-import giskardpy.identifier as identifier
 
 
 class VisualizationBehavior(GiskardBehavior):
-    def __init__(self, name):
-        super(VisualizationBehavior, self).__init__(name)
-        self.enable_visualization = self.get_god_map().safe_get_data(identifier.marker_visualization)
-
     def setup(self, timeout):
         self.publisher = rospy.Publisher(u'~visualization_marker_array', MarkerArray, queue_size=1)
         self.robot_base = self.get_robot().get_root()
         return super(VisualizationBehavior, self).setup(timeout)
 
     def update(self):
-        if not self.enable_visualization:
-            return py_trees.common.Status.SUCCESS
-
         markers = []
         robot = self.get_robot()
         get_fk = robot.get_fk_pose
@@ -36,7 +28,8 @@ class VisualizationBehavior(GiskardBehavior):
 
             marker.header.frame_id = self.robot_base
             marker.action = Marker.ADD
-            marker.id = int(hashlib.md5(link_name).hexdigest()[:6], 16) # FIXME find a better way to give the same link the same id
+            marker.id = int(hashlib.md5(link_name).hexdigest()[:6],
+                            16)  # FIXME find a better way to give the same link the same id
             marker.ns = u'planning_visualization'
             marker.header.stamp = rospy.Time()
